refactor(client): P5b(1/5) — scheduling handlers to native_tools/scheduling.py
Verbatim body moves: schedule_task/list/update/delete_schedule/parse_time
+ _validate_schedule_payload/_extract_tool_input_from_steps into
SchedulingTools. Bot keeps one-line delegates (dispatch host + test seam;
smoke tests call _validate_schedule_payload directly).
